definition of class Movable was taken to individual file
